The mod compiles and the game boots

+++ src/main/java/dev/jb0s/blockgameenhanced/event/network/ServerPingEvent.java
@@ -10,5 +10,5 @@ public interface ServerPingEvent {
         }
     });
 
-    void serverPing(int id);
+    void serverPing(long id);
 }
